fix(conversion): keep blockPath dep-free — extract mergeBlockValues to slateMerge
The bridge imported mergeBlockValues from slateTransforms.js, which imports
@plone/volto-slate and @plone/volto/registry. blockPath.js is bundled by the
offline block-path evaluator (esbuild — block-sanity, buildBlockPathMap), which
can't resolve those Volto-only deps, so discovery failed to build. mergeBlockValues
is a pure function, so move it to a dep-free slateMerge.js and import it there;
slateTransforms.js re-exports it for back-compat. Restores the DI boundary that
kept blockPath.js free of the registry/volto-slate imports.
